+++ rocprim/include/rocprim/device/device_run_length_encode_config.hpp
@@ -22,11 +22,12 @@
 #define ROCPRIM_DEVICE_DEVICE_RUN_LENGTH_ENCODE_CONFIG_HPP_
 
 #include "config_types.hpp"
-#include "detail/device_config_helper.hpp"
+#include "device_reduce_by_key_config.hpp"
 
 #include "../config.hpp"
 #include "../detail/various.hpp"
 #include "../type_traits.hpp"
+#include "detail/config/device_run_length_encode.hpp"
 #include "detail/config/device_run_length_encode_non_trivial.hpp"
 
 #include <type_traits>
@@ -42,10 +43,7 @@ BEGIN_ROCPRIM_NAMESPACE
 /// Must be \p reduce_by_key_config or \p default_config.
 /// \tparam SelectConfig - configuration of device-level select operation.
 /// Must be \p select_config or \p default_config.
-template<
-    class ReduceByKeyConfig,
-    class SelectConfig = default_config
->
+template<typename ReduceByKeyConfig, typename SelectConfig = default_config>
 struct run_length_encode_config
 {
     /// \brief Configuration of device-level reduce-by-key operation.
@@ -56,6 +54,58 @@ struct run_length_encode_config
 
 namespace detail
 {
+
+template<typename ReduceByKeyConfig,
+         typename KeyType,
+         typename AccumulatorType,
+         typename BinaryFunction>
+struct wrapped_trivial_runs_config
+    : wrapped_reduce_by_key_config<ReduceByKeyConfig, KeyType, AccumulatorType, BinaryFunction>
+{};
+
+template<typename ReduceByKeyConfig,
+         typename SelectConfig,
+         typename KeyType,
+         typename AccumulatorType,
+         typename BinaryFunction>
+struct wrapped_trivial_runs_config<
+    rocprim::run_length_encode_config<ReduceByKeyConfig, SelectConfig>,
+    KeyType,
+    AccumulatorType,
+    BinaryFunction>
+    : wrapped_reduce_by_key_config<ReduceByKeyConfig, KeyType, AccumulatorType, BinaryFunction>
+{};
+
+template<typename KeyType,
+         typename AccumulatorType,
+         typename BinaryFunction,
+         typename Enable = void>
+struct wrapped_trivial_runs_impl
+    : wrapped_reduce_by_key_impl<KeyType, AccumulatorType, BinaryFunction, Enable>
+{};
+
+template<typename KeyType, typename AccumulatorType, typename BinaryFunction>
+struct wrapped_trivial_runs_impl<
+    KeyType,
+    AccumulatorType,
+    BinaryFunction,
+    std::enable_if_t<is_arithmetic<KeyType>::value && is_arithmetic<AccumulatorType>::value
+                     && is_binary_functional<BinaryFunction>::value>>
+{
+    template<target_arch Arch>
+    struct architecture_config
+    {
+        static constexpr reduce_by_key_config_params params
+            = default_trivial_runs_config<static_cast<unsigned int>(Arch),
+                                          KeyType,
+                                          AccumulatorType>{};
+    };
+};
+
+template<typename KeyType, typename AccumulatorType, typename BinaryFunction>
+struct wrapped_trivial_runs_config<default_config, KeyType, AccumulatorType, BinaryFunction>
+    : wrapped_trivial_runs_impl<KeyType, AccumulatorType, BinaryFunction>
+{};
 
 // Wrap around run_length_encode_config and the newly added non_trivial_runs_config for the
 // run_length_encode_non_trivial_runs algorithm. Three cases are considered for selecting
@@ -84,7 +134,7 @@ struct wrapped_non_trivial_runs_config
     };
 };
 
-template<class ReduceByKeyConfig, class SelectConfig, class InputType>
+template<typename ReduceByKeyConfig, typename SelectConfig, typename InputType>
 struct wrapped_non_trivial_runs_config<
     rocprim::run_length_encode_config<ReduceByKeyConfig, SelectConfig>,
     InputType>
@@ -144,13 +194,23 @@ struct wrapped_non_trivial_runs_config<default_config, InputType>
 
 #ifndef DOXYGEN_DOCUMENTATION_BUILD
 
+template<typename KeyType, typename AccumulatorType, typename BinaryFunction>
+template<target_arch Arch>
+constexpr reduce_by_key_config_params wrapped_trivial_runs_impl<
+    KeyType,
+    AccumulatorType,
+    BinaryFunction,
+    std::enable_if_t<is_arithmetic<KeyType>::value && is_arithmetic<AccumulatorType>::value
+                     && is_binary_functional<BinaryFunction>::value>>::architecture_config<Arch>::
+    params;
+
 template<typename RLENonTrivialRunsConfig, typename InputType>
 template<target_arch Arch>
 constexpr non_trivial_runs_config_params
     wrapped_non_trivial_runs_config<RLENonTrivialRunsConfig,
                                     InputType>::architecture_config<Arch>::params;
 
-template<class ReduceByKeyConfig, class SelectConfig, typename InputType>
+template<typename ReduceByKeyConfig, typename SelectConfig, typename InputType>
 template<target_arch Arch>
 constexpr non_trivial_runs_config_params wrapped_non_trivial_runs_config<
     rocprim::run_length_encode_config<ReduceByKeyConfig, SelectConfig>,